bitcoin_charts: rename bitcoin_price to bitcoin_charts
